mobile: Move tests in main_interface_test.cc to internal_engine_test.cc (#32526)
